+++ scripts/mp7patch.py
@@ -31,23 +31,23 @@ def replace_area_constraints(filename):
         file.truncate()
         file.write( fileContents )
         
-#def replace_brd_decl(filename):
-    #content = tb.read_file(filename)
+def replace_brd_decl(filename):
+    content = tb.read_file(filename)
 
-    #expr_nregion = re.compile(r"(constant\s+N_REGION\s*:\s*integer\s*:=\s*)(\d+)(\s*;)")
-    #expr_crossregion = re.compile(r"(constant\s+CROSS_REGION\s*:\s*integer\s*:=\s*)(\d+)(\s*;)")
+    expr_nregion = re.compile(r"(constant\s+N_REGION\s*:\s*integer\s*:=\s*)(\d+)(\s*;)")
+    expr_crossregion = re.compile(r"(constant\s+CROSS_REGION\s*:\s*integer\s*:=\s*)(\d+)(\s*;)")
 
-    #content, count = expr_nregion.subn(r"\g<1>8\g<3>", content)
-    #if count != 1:
-        #raise RuntimeError("Could not replace the N_REGION value.")
+    content, count = expr_nregion.subn(r"\g<1>8\g<3>", content)
+    if count != 1:
+        raise RuntimeError("Could not replace the N_REGION value.")
 
-    #content, count = expr_crossregion.subn(r"\g<1>7\g<3>", content)
-    #if count != 1:
-        #raise RuntimeError("Could not replace the CROSS_REGION value.")
+    content, count = expr_crossregion.subn(r"\g<1>7\g<3>", content)
+    if count != 1:
+        raise RuntimeError("Could not replace the CROSS_REGION value.")
 
-    #with open(filename, "w") as fp:
-        #fp.write(content)
-        #logging.info("Successfully patched brd_decl file '{}'".format(filename))
+    with open(filename, "w") as fp:
+        fp.write(content)
+        logging.info("Successfully patched brd_decl file '{}'".format(filename))
 
 def insert_l1a_ttc(filename):
     content = tb.read_file(filename)
@@ -88,14 +88,14 @@ def patch_all(projectpath):
     """Batch patch all firmware files."""
     root_path = os.path.abspath(projectpath)
 
-    #brd_decl_path = 'boards/mp7/base_fw/mp7xe_690/firmware/hdl/mp7_brd_decl.vhd'
+    brd_decl_path = 'boards/mp7/base_fw/mp7xe_690/firmware/hdl/mp7_brd_decl.vhd'
     area_constraints_path = 'boards/mp7/base_fw/common/firmware/ucf/area_constraints.tcl'
     mp7xe_690_path = 'boards/mp7/base_fw/mp7xe_690/firmware/hdl/mp7xe_690.vhd'
     #script_writer_path = 'scripts/firmware/dep_tree/VivadoScriptWriter.py'
 
     #append_vivado_rules(os.path.join(root_path, script_writer_path))
     replace_area_constraints(os.path.join(root_path, area_constraints_path))
-    #replace_brd_decl(os.path.join(root_path, brd_decl_path))
+    replace_brd_decl(os.path.join(root_path, brd_decl_path))
     insert_l1a_ttc(os.path.join(root_path, mp7xe_690_path))
 
 def parse_args():