+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [v1.22.1] - 2023-04-07
+### Comment
+
+- mp7_ugt firmware release v1.22.1
+- based on v1.22.0, but without links 32..71
+
 ## [v1.22.0] - 2023-03-24
 ### Comment